Remove of TS code
